html/parsing/html5lib/html-tree/tests26.dat

#data
<!DOCTYPE html><body><a href='#1'><nobr>1<nobr></a><br><a href='#2'><nobr>2<nobr></a><br><a href='#3'><nobr>3<nobr></a>
#errors
nobr in nobr
reconstruct at <br>
reconstruct at <a>
nobr in nobr
nobr in nobr
reconstruct at <br>
reconstruct at <a>
nobr in nobr
nobr in nobr
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <a>
|       href="#1"
|       <nobr>
|         "1"
|       <nobr>
|     <nobr>
|       <br>
|       <a>
|         href="#2"
|     <a>
|       href="#2"
|       <nobr>
|         "2"
|       <nobr>
|     <nobr>
|       <br>
|       <a>
|         href="#3"
|     <a>
|       href="#3"
|       <nobr>
|         "3"
|       <nobr>

#data
<!DOCTYPE html><body><b><nobr>1<nobr></b><i><nobr>2<nobr></i>3
#errors
nobr in nobr
AAA at </b>
reconstruct at <i>
nobr in nobr
nobr in nobr
AAA at </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|       <nobr>
|     <nobr>
|       <i>
|     <i>
|       <nobr>
|         "2"
|       <nobr>
|     <nobr>
|       "3"

#data
<!DOCTYPE html><body><b><nobr>1<table><nobr></b><i><nobr>2<nobr></i>3
#errors
1:45:in table:nobr
1:49:in table:/:b
1:49:unmatched end tag:b
1:52:in table:i
1:58:in table:nobr
1:58:in nobr:nobr
1:58:in table:/:nobr
1:58:not closed:i
1:58:in table:nobr
1:59:in table:\#text
1:65:in table:nobr
1:65:in nobr:nobr
1:65:in table:/:nobr
1:65:in table:nobr
1:69:in table:/:i
1:69:not closed:nobr
1:69:in table:\#text
1:69:in body:\#eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|         <nobr>
|           <i>
|         <i>
|           <nobr>
|             "2"
|           <nobr>
|         <nobr>
|           "3"
|         <table>

#data
<!DOCTYPE html><body><b><nobr>1<table><tr><td><nobr></b><i><nobr>2<nobr></i>3
#errors
unmatched </b>
<nobr> in nobr
AAA at implied </nobr>
<nobr> in nobr
AAA at impied </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|         <table>
|           <tbody>
|             <tr>
|               <td>
|                 <nobr>
|                   <i>
|                 <i>
|                   <nobr>
|                     "2"
|                   <nobr>
|                 <nobr>
|                   "3"

#data
<!DOCTYPE html><body><b><nobr>1<div><nobr></b><i><nobr>2<nobr></i>3
#errors
1:43:in nobr:nobr
1:43:not closed:div
1:47:not closed:nobr
1:47:not closed:nobr
1:56:in nobr:nobr
1:56:not closed:i
1:63:in nobr:nobr
1:67:not closed:nobr
1:67:in body:\#eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|     <div>
|       <b>
|         <nobr>
|         <nobr>
|       <nobr>
|         <i>
|       <i>
|         <nobr>
|           "2"
|         <nobr>
|       <nobr>
|         "3"

#data
<!DOCTYPE html><body><b><nobr>1<nobr></b><div><i><nobr>2<nobr></i>3
#errors
1
2
3
4
5
6
7
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|       <nobr>
|     <div>
|       <nobr>
|         <i>
|       <i>
|         <nobr>
|           "2"
|         <nobr>
|       <nobr>
|         "3"

#data
<!DOCTYPE html><body><b><nobr>1<nobr><ins></b><i><nobr>
#errors
1
2
3
4
5
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|       <nobr>
|         <ins>
|     <nobr>
|       <i>
|     <i>
|       <nobr>

#data
<!DOCTYPE html><body><b><nobr>1<ins><nobr></b><i>2
#errors
1
2
3
4
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|         <ins>
|       <nobr>
|     <nobr>
|       <i>
|         "2"

#data
<!DOCTYPE html><body><b>1<nobr></b><i><nobr>2</i>
#errors
1
2
3
4
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <nobr>
|     <nobr>
|       <i>
|     <i>
|       <nobr>
|         "2"

#data
| <p><code x</code></p>
| 
#errors
AAA
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <code>
|         code=""
|         x<=""
|     <code>
|       code=""
|       x<=""
|       "
"

#data
<!DOCTYPE html><svg><foreignObject><p><i></p>a
#errors
AAA
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <p>
|           <i>
|         <i>
|           "a"

#data
<!DOCTYPE html><table><tr><td><svg><foreignObject><p><i></p>a
#errors
AAA
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg foreignObject>
|                 <p>
|                   <i>
|                 <i>
|                   "a"

#data
<!DOCTYPE html><math><mtext><p><i></p>a
#errors
AAA
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <p>
|           <i>
|         <i>
|           "a"

#data
<!DOCTYPE html><table><tr><td><math><mtext><p><i></p>a
#errors
AAA
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mtext>
|                 <p>
|                   <i>
|                 <i>
|                   "a"

#data
<!DOCTYPE html><body><div><!/div>a
#errors
29: Bogus comment.
34: End of file seen and there were open elements.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       <!-- /div -->
|       "a"
